move mmap.jl into util.jl, move & add tests to runtests.jl, fixup things
