+++ src/JuliaDB.jl
@@ -7,7 +7,6 @@ export NDSparse, Columns
 
 include("util.jl")
 include("interval.jl")
-include("mmap.jl")
 include("dtable.jl")
 include("load.jl")
 include("load-save.jl")
